requirements.txt
torch>=1.10.0
torchvision>=0.11.0
Pillow>=8.0.0
numpy>=1.19.0
tqdm>=4.50.0
matplotlib>=3.3.0
pandas>=1.2.0
scikit-learn>=0.24.0
jupyter>=1.0.0
pytest>=6.0.0
